Add files via upload
This is dataset we have obtained from online repository known as UCI Machine Learning Repository. It is a collection of databases, domain theories, and data generators that are used by the machine learning community for the empirical analysis of machine learning algorithms.
The file format for dataset is .csv (comma separated values) which contains around 195 samples of patients.